+++ packages/admin/dashboard/src/dashboard-app/routes/get-route.map.tsx
@@ -8,6 +8,7 @@ import { PublicLayout } from "../../components/layout/public-layout"
 import { SettingsLayout } from "../../components/layout/settings-layout"
 import { ErrorBoundary } from "../../components/utilities/error-boundary"
 import { PermissionsRequirementsProvider } from "../../providers/permissions-provider"
+import { HydrateFallback } from "../../components/utilities/hydrate-fallback"
 
 export function getRouteMap({
   settingsRoutes,
@@ -20,6 +21,7 @@ export function getRouteMap({
     {
       element: <ProtectedRoute />,
       errorElement: <ErrorBoundary />,
+      hydrateFallbackElement: <HydrateFallback />,
       children: [
         {
           element: <MainLayout />,
@@ -1361,6 +1363,7 @@ export function getRouteMap({
     {
       element: <ProtectedRoute />,
       errorElement: <ErrorBoundary />,
+      hydrateFallbackElement: <HydrateFallback />,
       children: [
         {
           path: "/settings",