+++ packages/core/core-flows/src/cart/workflows/get-variants-and-items-with-prices.ts
@@ -29,6 +29,7 @@ import { useQueryGraphStep } from "../../common"
 import { getVariantPriceSetsStep } from "../steps"
 import {
   cartFieldsForPricingContext,
+  productVariantsCacheTags,
   productVariantsFields,
 } from "../utils/fields"
 import {
@@ -114,8 +115,7 @@ export const prepareVariantsAndItemsWithPricesStep = createStep(
         calculatedPriceSet = calculatedPriceSets[item_.variant_id!]
       }
 
-      const isCustomPrice =
-        item_.is_custom_price ?? isDefined(item?.unit_price)
+      const isCustomPrice = item_.is_custom_price ?? isDefined(item?.unit_price)
 
       if (!calculatedPriceSet && item_.variant_id && !isCustomPrice) {
         priceNotFound.push(item_.variant_id)
@@ -209,20 +209,19 @@ export const getVariantsAndItemsWithPrices = createWorkflow(
   (
     input: WorkflowData<GetVariantsAndItemsWithPricesWorkflowInput>
   ): WorkflowResponse<GetVariantsAndItemsWithPricesWorkflowOutput> => {
-    const variantIds = transform(
-      { input },
-      (data): string[] => {
-        if (data.input.variants?.id) {
-          return data.input.variants.id
-        }
-
-        return Array.from(
-          new Set(
-            (data.input.cart.items ?? data.input.items ?? []).map((i) => i.variant_id)
-          )
-        ).filter((v): v is string => !!v)
+    const variantIds = transform({ input }, (data): string[] => {
+      if (data.input.variants?.id) {
+        return data.input.variants.id
       }
-    )
+
+      return Array.from(
+        new Set(
+          (data.input.cart.items ?? data.input.items ?? []).map(
+            (i) => i.variant_id
+          )
+        )
+      ).filter((v): v is string => !!v)
+    })
 
     const cartPricingContext = transform(
       {
@@ -281,7 +280,8 @@ export const getVariantsAndItemsWithPrices = createWorkflow(
       },
       options: {
         cache: {
-          enable: true,
+          tags: productVariantsCacheTags,
+          computeAutomaticTags: true,
         },
       },
     }).config({ name: "fetch-variants" })